FIX: style event post header
