+++ src/app/blog/page.tsx
@@ -25,7 +25,7 @@ import { type NextServerSearchParams } from '@/types/searchParams'
 import { getCategorySettings } from '@/utils/categoryUtils'
 import { createMetadata } from '@/utils/createMetadata'
 import { getBlogPostsData } from '@/utils/getBlogPostData'
-import { getMetaData } from '@/utils/getMetaData'
+import { getBlogPostMetaData } from '@/utils/getMetaData'
 
 import { attributes } from '@/content/pages/blog.md'
 
@@ -90,7 +90,7 @@ export default function Blog({ searchParams }: Props) {
         isFeatured
         title={featuredPost.title}
         description={featuredPost.description}
-        metaData={getMetaData(featuredPost.publishedOn)}
+        metaData={getBlogPostMetaData(featuredPost.publishedOn)}
         image={{
           type: 'dynamic',
           ...featuredPost.image,
@@ -161,7 +161,7 @@ export default function Blog({ searchParams }: Props) {
                           description={description}
                           image={image}
                           textIsClamped={true}
-                          metaData={getMetaData(publishedOn)}
+                          metaData={getBlogPostMetaData(publishedOn)}
                           cta={{
                             href: `${PATHS.BLOG.path}/${slug}`,
                             text: 'Read Post',
